+++ tests/testthat/test-install-dev.R
@@ -1,5 +1,3 @@
-context("test-install-dev")
-
 test_that("install_dev works with GitHub URLs", {
   skip_on_cran()
   skip_if_offline()
@@ -39,15 +37,6 @@ test_that("install_dev fails if there is no URL field with a GitHub, GitLab or B
   skip_if_offline()
 
   expect_error(install_dev("XML"), "Could not determine development repository")
-})
-
-test_that("install_dev works with GitLab URLs", {
-  skip_on_cran()
-  skip_if_offline()
-
-  mockery::stub(install_dev, "install_gitlab", identity)
-
-  expect_equal(install_dev("iemiscdata"), "iembry/iemiscdata")
 })
 
 test_that("install_dev works with Bitbucket URLs", {