+++ tests/testthat/test-install-git.R
@@ -1,6 +1,3 @@
-
-context("Install from git repo")
-
 test_that("install_git with git2r", {
 
   skip_on_cran()
@@ -29,66 +26,6 @@ test_that("install_git with git2r", {
   expect_equal(remote$url, url)
   expect_equal(remote$ref, NULL)
   expect_equal(remote_sha(remote), remote$sha)
-  expect_true(!is.na(remote$sha) && nzchar(remote$sha))
-})
-
-test_that("install_git with git2r and ref", {
-
-  skip_on_cran()
-  skip_if_offline()
-  skip_if_over_rate_limit()
-  skip_if_not_installed("git2r")
-
-  Sys.unsetenv("R_TESTS")
-
-  lib <- tempfile()
-  on.exit(unlink(lib, recursive = TRUE), add = TRUE)
-  dir.create(lib)
-
-  url <- "https://github.com/gaborcsardi/pkgconfig.git"
-  install_git(url, lib = lib, ref = "travis", git = "git2r", quiet = TRUE)
-
-  expect_silent(packageDescription("pkgconfig", lib.loc = lib))
-  expect_equal(
-    packageDescription("pkgconfig", lib.loc = lib)$RemoteUrl,
-    url
-  )
-
-  remote <- package2remote("pkgconfig", lib = lib)
-  expect_s3_class(remote, "remote")
-  expect_s3_class(remote, "git2r_remote")
-  expect_equal(format(remote), "Git")
-  expect_equal(remote$url, url)
-  expect_equal(remote$ref, "travis")
-  expect_equal(remote_sha(remote), remote$sha)
-  expect_true(!is.na(remote$sha) && nzchar(remote$sha))
-})
-
-
-test_that("install_git with command line git", {
-
-  skip_on_cran()
-  skip_if_offline()
-  if (is.null(git_path())) skip("git is not installed")
-
-  Sys.unsetenv("R_TESTS")
-
-  lib <- tempfile()
-  on.exit(unlink(lib, recursive = TRUE), add = TRUE)
-  dir.create(lib)
-
-  url <- "https://github.com/cran/falsy.git"
-  install_git(url, git = "external", lib = lib, quiet = TRUE)
-
-  expect_silent(packageDescription("falsy", lib.loc = lib))
-  expect_equal(packageDescription("falsy", lib.loc = lib)$RemoteUrl, url)
-
-  remote <- package2remote("falsy", lib = lib)
-  expect_s3_class(remote, "remote")
-  expect_s3_class(remote, "xgit_remote")
-  expect_equal(format(remote), "Git")
-  expect_equal(remote$url, url)
-  expect_equal(remote$ref, NULL)
   expect_true(!is.na(remote$sha) && nzchar(remote$sha))
 })
 
@@ -203,7 +140,7 @@ test_that("remote_package_name.git2r_remote returns the package name if it exist
   expect_equal(remote_package_name(remote), "igraph")
 
   # works with tags in URL (and different name from repo name)
-  url <- "https://github.com/igraph/rigraph.git@master"
+  url <- "https://github.com/igraph/rigraph.git@main"
   remote <- git_remote(url, git = "git2r")
   expect_equal(remote_package_name(remote), "igraph")
 
@@ -211,17 +148,6 @@ test_that("remote_package_name.git2r_remote returns the package name if it exist
   url <- "https://gitlab.com/r-lib-grp/test-pkg.git"
   remote <- git_remote(url, git = "git2r")
   expect_equal(remote_package_name(remote), "test123")
-
-  # safely returns NA when DESCRIPTION url is not accessible
-  # (condition emitted due to inaccessible git remote for remote_sha during testing)
-  url <- "https://gitlab.com/r-lib-grp/fake-private-repo.git"
-  remote <- git_remote(url, git = "git2r")
-  err <- tryCatch(remote_sha(remote), error = function(e) e)
-  expect_error(  # expect same error as calling remote_sha directly
-    expect_equal(remote_package_name(remote), NA_character_),
-    class = class(err),
-    label = conditionMessage(err)
-  )
 })
 
 test_that("remote_package_name.xgit_remote returns the package name if it exists", {
@@ -241,26 +167,9 @@ test_that("remote_package_name.xgit_remote returns the package name if it exists
   expect_equal(remote_package_name(remote), "igraph")
 
   # works with tags in URL (and different name from repo name)
-  url <- "https://github.com/igraph/rigraph.git@master"
+  url <- "https://github.com/igraph/rigraph.git@main"
   remote <- git_remote(url, git = "external")
   expect_equal(remote_package_name(remote), "igraph")
-})
-
-test_that("remote_sha.xgit remote returns the SHA if it exists", {
-
-  skip_on_cran()
-  skip_if_offline()
-  if (is.null(git_path())) skip("git is not installed")
-
-  url <- "https://github.com/cran/falsy.git"
-
-  # works with tags
-  remote <- git_remote(url, ref = "1.0", git = "external")
-  expect_equal(remote_sha(remote), "0f39d9eb735bf16909831c0bb129063dda388375")
-
-  # works with full SHAs
-  remote <- git_remote(url, ref = "26a36cf957a18569e311ef75b6f61f822de945ef", git = "external")
-  expect_equal(remote_sha(remote), "26a36cf957a18569e311ef75b6f61f822de945ef")
 })
 
 test_that("remote_metadata.xgit_remote", {